backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Azure SDK
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-servicebus==7.11.4
azure-search-documents==11.4.0
azure-monitor-opentelemetry==1.1.0

# AI and ML
openai==1.3.5
azure-openai==1.0.0b3
numpy==1.24.4
tiktoken==0.5.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
ed25519==1.5

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.4
pyyaml==6.0.1
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and logging
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
